Remove parentheses: add semicolon before lambda if necessary
In 0801eaa7, it was suggested just not to suggest parentheses removal.
However, it contradicts with behaviour of "Replace contains" which
adds semicolon before lambda in this case. So in this commit I made
behaviour of remove parentheses the same.

This commit reverts production part of 0801eaa7
Fixes failing test of "replace contains"

GitOrigin-RevId: 48236f9424688174e95a22cd26c62249cec0f852