FIR: Refactor designated body transformers
The idea is making their mechanics more transparent and clear
Namely, we should move to the next level of designation only for declaration's content

Otherwise there might be some wrong steps when somebody just randomly
calls "transformElement" on the parent declaration.
For example, it happens for local classes when we resolve their supertypes, etc.

GitOrigin-RevId: 5378d6b92527d6500d8eedb6b8ca5835f8ceab3c