Changed gradle-plugin version in Gradle tests.
When 6.5.1 Gradle version was added, more tests have began to fail
because of legacy gradle-plugin versions (1.1., 1.2.) which are not
compatible with the new version of Gradle. This commit updates the used
version of the Gradle plugin to the latest stable (1.3.72) and makes it
configurable from tests.

Original commit: 05d6217f78e1f691fcda1a27fb17a7d87800527e

GitOrigin-RevId: 4d7d64421183fc3e07bf20e86198858a918592f4