[JVM_IR] Fix offsets in constant propagation optimization.
Loads of temporary variables that contain constants are replaced
with a copy of the constant. This avoids locals loads and stores.
However, the copy of the constant needs to have the offset of
the load and not of the original constant.

Fixes KT-41963.

Original commit: 78483930bc259d12fbe1217dfa736b17aefb28b5

GitOrigin-RevId: 9bd03a549a2e304cb9d28d7cc88ce8a12214a04b